Fix configuration via code (#78)
Related to #73. The DefaultInstance would be initialized with the
default Config eagerly when building a new instance with configuration
via code. If there were no URL environment variable set, the
DefaultInstance init would throw due to a missing URL in its config.

Fix by ensuring that the default Config is initialized lazily.