Remove DefaultInstance (#86)
Removing `DefaultInstance` further cleans up the library API so that its
entrypoint is more familiar and easier to grasp with autocomplete. It's
not much effort to build an instance before using.

Also make the `check` task run integration tests and examples, and
switch CI to run `test` instead. CI can't run integration or examples
because it doesn't have an API token. Also fix the `update-examples`
workflow to support `workflow_dispatch`.