Declare library version in gradle.properties (#175)
The release flow could be more standard. Currently:

1. version on source code is always `SNAPSHOT`
2. Tag is pushed, e.g. `2023.4.0` (`version=SNAPSHOT` even on tag ref)
3. A `publish` build is ran on tag ref to publish the artifact passing
`-Pversion=2023.4.0`
4. The README is only then updated on main for the new version (on tag
ref, the README is obsolete)

But in most OSS projects, the version on source code corresponds to the
version that code was released in, e.g. `version=2023.3.0` on the
`2023.3.0` tag ref, and `main` declares the next version,
`version=2023.4.0` ("Prepare for next release" commits). Adjusting to
this also makes README versioning more intuitive.