Add tests on generated API post-processing (#181)
Extract the logic of post-processing the generated Develocity API code
into a proper Task and add tests. Allows it to be tested on its own, as
the test setup for the whole plugin would be more complex (arguably not
worth it for such a small project). Setup could be simpler by not using
Ant and just unit-testing it, but simply having tests is a more
important first step.

Also remove the "Workaround for properties generated with
`arrayListOf(null,null)` as default value". It was no longer needed
(fixed in some past openapi-generator release), hence the resulting API
is unchanged (see PR comment with diff).

On `generated-api-diff` the changes for base-ref-checkout must be done
in a separate PR for diffing to work for this PR. In base ref, the task
doesn't exist yet and the output dir is the old one.