Restore compatibility with older Kotlin versions (#434)
Fixes compatibility with Kotlin/Jupyter versions 0.12.0.322 and older*
(tracked in #426) by downgrading the language version the library is
compiled for to 1.8.

The library was previously compiled for 1.9, which was compatible with
these Kotlin/Jupyter versions, then upgraded to Kotlin Gradle plugin
2.x, causing Kotlin/Jupyter versions still on compiler 1.9.x to no
longer support running the library. This patch broadens compatibility
beyond what was initially supported by downgrading to language version
1.8.

Also addresses compatibility with projects using the library as part of
a Gradle plugin, in which the embedded Kotlin compiler version in Gradle
>=8.2 will only support libraries compiled for language version >=1.8.
Reported in #404.

*Further Kotlin/Jupyter versions between 0.12.0.322 and 0.14.1.514 are
probably incompatible as well, depending on when Kotlin/Jupyter upgraded
its compiler version to 2.x.