Run integration tests and examples in CI (#435)
Run `integrationTest` and `examplesTest` test suites in CI as part of PR
checks and before publishing artifacts, as a final validation.

#432 and this PR are aimed at preventing regressions such as #426.

These test suites require a real Develocity server to make API requests
to. This was previously not possible because the maintainer didn't have
access to a Develocity server that could be used from a public GitHub
repository's workflows. As a result, the `integrationTest` suite and the
examples had to be manually ran locally against the Develocity server
that was available. That was always done for significant PRs and before
each release. Now, with access to a server that can be used in this
repository, changes can be fully validated in CI, reducing the effort
needed to make changes to the library.

Also fix the `examplesTest` notebooks test which broke from the
`@TempDir` annotation being accidentally removed in
309acfe87f4b2381d32576eef3e5a4b39c7b44d8. Since it was a trivial change,
tests were not ran locally. After this PR, a commit such as this being
merged will no longer be possible.