Improve running integration tests and examples in CI (#437)
- Re-enable output capturing in build scans

For the same reasons stated in #435, output capturing had to be disabled
to avoid leaking information about the Develocity server the library was
tested with locally. With the current Develocity server it's being
tested with, that's no longer a concern. Having console and tests output
visible in build scans will make troubleshooting CI failures
significantly easier.

- Allow running examples in parallel, which should make CI faster
- Enforce cache is disabled when running `:examples` tasks to guarantee
that it is making the expected requests to Develocity (was already
disabled for `examplesTest` test suite). There are pros and cons to
this, but currently safety is preferred.

All aimed at making CI faster and easier to troubleshoot, after #435.
Part of work done to prevent regressions such as #426.