Test examples notebooks with library descriptor (#439)
Improve `NotebooksTest` to also test the Jupyter integration. Before
this change, by using Kotlin scripting `DependsOn` and imports, the
library's integration with Kotlin Jupyter was not tested. After this
change, by using the `%use` magic line, Kotlin Jupyter will use the
library's integration as part of the test. Currently, that integration
is merely importing some classes, but this change is still an
improvement in that it increases coverage.