Bump okhttp to 5.1.0 (#440)
Bump okhttp to 5.1.0 and re-order dependency declarations so that the
Kotlin kernel for Jupyter will use the higher version of `okio`,
`3.15.0`, which is requested and required by `okhttp:5.1.0`, avoiding a
runtime failure.

The change in the buildscript is reflected in the library's published
`.pom` file. It appears that the kernel will pick the first seen version
of an artifact when traversing the dependencies and does not consider
dependency constraints. `okhttp` must be declared first in the pom file
and `moshi`, which depends on an older version of `okio`, must be
declared later.

This bump without re-ordering dependencies would result in a failure as
described in [Kotlin/kotlin-jupyter#484][issue] and seen in this [build
scan][failing-build] from #425.

[issue]: https://github.com/Kotlin/kotlin-jupyter/issues/484
[failing-build]:
https://scans.gradle.com/s/3mvhfatr7gduy/tests/task/:library:examplesTest/details/com.gabrielfeo.develocity.api.example.notebook.NotebooksTest/testMostFrequentBuildsNotebook()?focused-execution=1&page=eyJvdXRwdXQiOnsiMCI6NH19&top-execution=1#L795