Support conventional access key sources (#455)
Support using an access key from the conventional locations, matching
the official tooling. Simplifies configuration for users of this
library, which most likely also use the official tooling. Currently,
they're required to set both `DEVELOCITY_API_TOKEN` (for this library)
and `DEVELOCITY_ACCESS_KEY` (for official tooling), which is possibly
redundant if the access keys are from the same user and instance. If an
access key different from the one defined in usual locations must be
used for the API, then a custom `Config.accessKey` is still supported.
The name "API token" was also unintuitive considering that all
Develocity documentation refers to them as "access keys". Resolves #338.

This is a breaking change for projects using `DEVELOCITY_API_TOKEN` or
`Config.apiToken`, targeting the upcoming 2025.1.0 major release.

Thanks to @mcumings for providing an initial reference implementation.