Support conventional server URL variable (#468)
As follow-up to #455 / #338, support the conventional server URL
variable that's already set up for official Develocity tooling, such as

- any Gradle project using [`common-custom-user-data-gradle-plugin`][1]
with a set [Develocity URL override][2]
- the [Develocity Python agent][3]

`Config.apiUrl` and `DEVELOCITY_API_URL` are renamed to `Config.server`
and `DEVELOCITY_URL` for consistency with official Develocity tooling
nomenclature.

`Config.server` is strongly typed as a URI and validation is improved.

[1]: https://github.com/gradle/common-custom-user-data-gradle-plugin
[2]:
https://github.com/gradle/common-custom-user-data-gradle-plugin/tree/main?tab=readme-ov-file#configuration-overrides
[3]:
https://docs.gradle.com/develocity/python-agent/#environment_variables