Change logLevel to not touch `simple-logger` `defaultLogLevel` (#368)
Fix #331

> `Config.logLevel` / `DEVELOCITY_API_LOG_LEVEL` internally changes
`org.slf4j.simpleLogger.defaultLogLevel`, which changes level of all
loggers, not only the library's as the javadoc says. In practice, this
is an issue when the user is using `simple-logger` and using loggers (or
using other dependencies that use loggers), as changing
`Config.logLevel` / `DEVELOCITY_API_LOG_LEVEL` will unexpectedly affect
those loggers.